requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
django-cors-headers==4.3.1
python-dotenv==1.0.0
Pillow==10.1.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
django-extensions==3.2.3
factory-boy==3.3.0
Faker==20.1.0